Moved DTOs to submodule
