+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "src/dtos"]
+	path = src/dtos
+	url = https://github.com/Benedict-Scrumberbatch/peer-recognition-dtos.git
