+++ src/app.module.ts
@@ -6,9 +6,9 @@ import { UsersModule } from './users/users.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { CompanyModule } from './company/company.module';
 import { RecognitionModule } from './recognition/recognition.module'
-import { Users } from './entity/users.entity';
-import { Tag } from './entity/tag.entity';
-import { Company } from './entity/company.entity';
+import { Users } from './dtos/entity/users.entity';
+import { Tag } from './dtos/entity/tag.entity';
+import { Company } from './dtos/entity/company.entity';
 import { RecognitionService } from './recognition/recognition.service';
 
 @Module({
