+++ src/company/company.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { CompanyService } from './company.service';
 import { CompanyController } from './company.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Company } from '../entity/company.entity';
+import { Company } from '../dtos/entity/company.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Company])],
