--- src/recognition/dto/create-rec.dto.ts
@@ -1,7 +1,0 @@
-export class CreateRecDto{
-    readonly company: number;
-    readonly employeeFrom: number;
-    readonly employeeTo: number;
-    readonly msg: string;
-    readonly tags: number[];
-}
