+++ src/recognition/recognition.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Post, Delete, Body, Param} from '@nestjs/common';
-import {CreateRecDto} from './dto/create-rec.dto'
+import {CreateRecDto} from '../dtos/dto/create-rec.dto'
 import {RecognitionService} from './recognition.service'
-import {Recognition} from '../entity/recognition.entity';
+import {Recognition} from '../dtos/entity/recognition.entity';
 
 @Controller('recognitions')
 export class RecognitionController {
