+++ src/users/users.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersController } from './users.controller';
-import { Users } from '../entity/users.entity';
-import { Login } from '../entity/login.entity';
-import { Company } from '../entity/company.entity';
-import { TagStats } from '../entity/tagstats.entity';
+import { Users } from '../dtos/entity/users.entity';
+import { Login } from '../dtos/entity/login.entity';
+import { Company } from '../dtos/entity/company.entity';
+import { TagStats } from '../dtos/entity/tagstats.entity';
 import { CompanyModule } from 'src/company/company.module';
-import { Recognition} from '../entity/recognition.entity';
+import { Recognition} from '../dtos/entity/recognition.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Users, Login, Company, Recognition, TagStats]), CompanyModule],
